+++ chatlearn/algorithm/grpo_utils/megatron_utils/train_helper.py
@@ -25,8 +25,26 @@ from megatron.training.utils import average_losses_across_data_parallel_group, g
 
 from chatlearn.utils import to_device
 
-from tokenizer import get_tokenizer
-from .utils import pad_to_max_len, generate_loss_mask_position_ids
+from .tokenizer_utils import get_tokenizer
+
+def pad_to_max_len(all_tokens_right_padded, max_len, pad_value):
+    pad_length = max_len - all_tokens_right_padded.size(1)
+    if pad_length <= 0:
+        return all_tokens_right_padded
+    # Pad the tensor with zeros on the right side to the desired length
+    padded_tensor = torch.nn.functional.pad(all_tokens_right_padded, (0, pad_length), mode='constant', value=pad_value)
+    return padded_tensor
+
+def generate_loss_mask_position_ids(tokens: torch.Tensor, prompt_token_length: list, response_token_length:list):
+    # Setup attention mask by prompt token length and response token length
+    loss_mask = torch.zeros_like(tokens, dtype=torch.int32, device=tokens.device)
+    for i in range(len(prompt_token_length)):
+        loss_mask[i, prompt_token_length[i]: prompt_token_length[i] + response_token_length[i]] = 1.0
+    _, seq_len = tokens.size()
+    position_ids = torch.arange(seq_len, dtype=torch.long, device=tokens.device)
+    position_ids = position_ids.unsqueeze(0).expand_as(tokens)
+
+    return loss_mask, position_ids
 
 def training_log(loss_dict, total_loss_dict, learning_rate, iteration,
                  loss_scale, report_memory_flag, skipped_iter,
