--- examples/mcore/configs/grpo/base.yaml
@@ -1,37 +1,0 @@
-includes:
-        - log.yaml
-
-seed: 1234
-tokenizer_type: ${tokenizer_type:NullTokenizer}
-patch_tokenizer_type: ${patch_tokenizer_type:NullTokenizer}
-tokenizer_model: ${tokenizer_model:"None"}
-vocab_file: ${vocab_file:"None"}
-merge_file: ${merge_file:"None"}
-vocab_size: ${vocab_size:32000}
-extra_vocab_size: ${extra_vocab_size:421}
-make_vocab_size_divisible_by: ${make_vocab_size_divisible_by:128}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
