--- examples/mcore/configs/grpo/log.yaml
@@ -1,11 +1,0 @@
-log_dir: ${log_dir:"None"}
-exp_name: ${exp_name:test}
-tensorboard_dir: ${tensorboard_dir:"None"}
-enable_tensorboard: ${enable_tensorboard:False}
-
-enable_wandb: ${enable_wandb:False}
-wandb_dir: ${wandb_dir}
-wandb_project: ${wandb_project}
-wandb_id: ${exp_name}
-wandb_name: ${exp_name}
-wandb_resume: ${WANDB_RESUME:allow}
