--- examples/mcore/configs/grpo/model_qwen3.yaml
@@ -1,24 +1,0 @@
-attention_dropout: ${attention_dropout:0.0}
-hidden_dropout: ${hidden_dropout:0.0}
-num_layers: ${policy_num_layers:28}
-hidden_size: ${policy_hidden_size:3584}
-num_attention_heads: ${policy_num_attention_heads:28}
-ffn_hidden_size: ${policy_ffn_hidden_size:18944}
-num_query_groups: ${policy_num_query_groups:4}
-seq_length: ${seq_length:2048}
-max_position_embeddings: ${max_position_embeddings:131072}
-swiglu: True
-normalization: ${normalization:RMSNorm}
-norm_epsilon: ${RMS_NORM_EPS:1e-6}
-use_rotary_position_embeddings: True
-position_embedding_type: ${position_embedding_type:rope}
-add_qkv_bias: false
-add_bias_linear: false
-qk_layernorm: true
-kv_channels: 128
-rotary_percent: 1.0
-rotary_base: 1000000
-rotary_seq_len_interpolation_factor: 1
-group_query_attention: True
-use_legacy_models: false
-untie_embeddings_and_output_weights: True
